client/requirements.txt
opencv-python==4.8.0.76
numpy==1.25.2
pillow==10.0.0
ttkbootstrap==1.10.1